added effect shadow toJson
