--- __TESTS__/unit/toJson/levelEffect.test.ts
@@ -1,12 +1,0 @@
-import {Transformation} from "../../../src";
-import {Effect} from "../../../src/actions/effect";
-
-describe('LevelEffect.toJson()', () => {
-  it('effect.sepia', () => {
-    const transformation = new Transformation()
-      .addAction(Effect.sepia());
-    expect(transformation.toJson()).toStrictEqual( [
-      { actionType: 'sepia' }
-    ]);
-  });
-});
