+++ src/actions/effect/EffectActions/SimpleEffectAction.ts
@@ -3,7 +3,7 @@ import {QualifierValue} from "../../../internal/qualifier/QualifierValue.js";
 import {Qualifier} from "../../../internal/qualifier/Qualifier.js";
 import {ExpressionQualifier} from "../../../qualifiers/expression/ExpressionQualifier.js";
 import {EFFECT_MODE_TO_ACTION_TYPE_MAP} from "../../../internal/internalConstants.js";
-import {ISimpleEffectActionModel} from "../../../internal/models/ISimpleEffectActionModel.js";
+import {ISimpleEffectActionModel} from "../../../internal/models/IEffectActionModel.js";
 
 /**
  * @description A class that defines a simple effect of the type e_{effectName}
