+++ src/actions/effect/Shadow.ts
@@ -4,6 +4,7 @@ import {Qualifier} from "../../internal/qualifier/Qualifier.js";
 import {prepareColor} from "../../internal/utils/prepareColor.js";
 import {SystemColors} from "../../qualifiers/color.js";
 import {ExpressionQualifier} from "../../qualifiers/expression/ExpressionQualifier.js";
+import {IShadowEffectActionModel} from "../../internal/models/IEffectActionModel";
 
 /**
  * @description Applies a shadow filter to the asset.
@@ -12,11 +13,12 @@ import {ExpressionQualifier} from "../../qualifiers/expression/ExpressionQualifi
  * @see Visit {@link Actions.Effect|Effect} for an example
  */
 class ShadowEffectAction extends Action {
+  protected _actionModel: IShadowEffectActionModel = {};
   private effectType: string;
 
   constructor(effectType: string, strength: number) {
     super();
-
+    this._actionModel.actionType = effectType;
     this.effectType = effectType;
     this.addQualifier(new Qualifier('e', new QualifierValue(['shadow', strength])));
   }
@@ -27,6 +29,7 @@ class ShadowEffectAction extends Action {
    * @return {this}
    */
   strength(strength: number): this {
+    this._actionModel.strength = strength;
     return this.addQualifier(new Qualifier('e', new QualifierValue(['shadow', strength])));
   }
 
@@ -36,6 +39,7 @@ class ShadowEffectAction extends Action {
    * @return {this}
    */
   offsetX(x:number | ExpressionQualifier): this {
+    this._actionModel.offsetX = x as number;
     return this.addQualifier(new Qualifier('x', new QualifierValue(x)));
   }
 
@@ -45,6 +49,7 @@ class ShadowEffectAction extends Action {
    * @return {this}
    */
   offsetY(y:number | ExpressionQualifier): this {
+    this._actionModel.offsetY = y as number;
     return this.addQualifier(new Qualifier('y', new QualifierValue(y)));
   }
 
@@ -54,6 +59,7 @@ class ShadowEffectAction extends Action {
    * @return {this}
    */
   color(color:SystemColors): this {
+    this._actionModel.color = color;
     return this.addQualifier(new Qualifier('co', new QualifierValue(prepareColor(color))));
   }
 }
