--- src/internal/models/IEffectActionWithLevelModel.ts
@@ -1,7 +1,0 @@
-import {IActionModel} from "./IActionModel.js";
-
-interface IEffectActionWithLevelModel extends IActionModel{
-  level?: number;
-}
-
-export {IEffectActionWithLevelModel};
