--- src/internal/models/ISimpleEffectActionModel.ts
@@ -1,7 +1,0 @@
-import {IActionModel} from "./IActionModel.js";
-
-interface ISimpleEffectActionModel extends IActionModel{
-
-}
-
-export {ISimpleEffectActionModel};
